better error codes implemented

+++ src/teams/teams.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { PrismaService } from 'src/prisma.service';
 import { TeamResponseDto } from './dto/team-response.dto';
 import { plainToInstance } from 'class-transformer';
@@ -57,7 +57,7 @@ export class TeamsService {
     return paginate(data, total, page, limit);
   }
 
-  async findOne(id: string, year: number): Promise<TeamResponseDto | null> {
+  async findOne(id: string, year: number): Promise<TeamResponseDto> {
     const team = await this.prisma.team.findUnique({
       where: {
         yearID_teamID: {
@@ -66,6 +66,9 @@ export class TeamsService {
         },
       },
     });
+    if (!team) {
+      throw new NotFoundException('Team not found');
+    }
 
     // Calculate stats using stat service
     const [stats] = await Promise.all([
@@ -84,7 +87,7 @@ export class TeamsService {
       }),
     ]);
     if (!stats) {
-      return this.mapToDto(team!);
+      return this.mapToDto(team);
     }
     const s = stats._sum;
 
@@ -113,10 +116,14 @@ export class TeamsService {
     };
 
     // Merge stats into team object for DTO mapping
-    return this.mapToDto(team!, career_batting);
+    return this.mapToDto(team, career_batting);
   }
 
   async update(id: number, updateTeamDto: UpdateTeamDto) {
+    const team = await this.prisma.team.findUnique({ where: { id } });
+    if (!team) {
+      throw new NotFoundException('Team not found');
+    }
     return await this.prisma.team.update({
       where: { id },
       data: updateTeamDto,
@@ -130,6 +137,10 @@ export class TeamsService {
   }
 
   async remove(id: number) {
+    const team = await this.prisma.team.findUnique({ where: { id } });
+    if (!team) {
+      throw new NotFoundException('Team not found');
+    }
     return await this.prisma.team.delete({
       where: { id },
     });
